requirements.txt
asgiref==3.8.1
boto3==1.35.43
botocore==1.35.43
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.10
jmespath==0.10.0
pyasn1==0.6.1
pydantic==1.10.18
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-jose==3.3.0
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
starlette==0.19.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.17.6
